requirements.txt
streamlit-authenticator==0.3.2
streamlit==1.32.0
typing_extensions
PyYAML